Update RefreshDatabaseWorker.kt:
if the database is empty, add two campaigns to the DB, add the participants from DataSource.kt to the DB, create and add the campaignParticpants to the DB

Update DAO's so id's are returned

+++ app/src/main/java/be/howest/ti/nma/dungeonsanddragonsinitiativetracker/data/db/entities/Campaign.kt
@@ -9,7 +9,7 @@ import androidx.room.PrimaryKey
     tableName = "campaigns"
 )
 data class Campaign(
-    @PrimaryKey(autoGenerate = true) val campaignId: Int = 0,
+    @PrimaryKey(autoGenerate = true) val campaignId: Long = 0,
     @DrawableRes val campaignImageDrawable: Int? = null,
     var campaignImageUri: Uri? = null,
     val campaignName: String,
